[#43] restructure events transport util
