+++ utils/io/logger.ts
@@ -1,0 +1,5 @@
+import { LoggerParams } from "./types";
+
+export const logger = (params: LoggerParams) => {
+  console.info(`[ApiConnection logger]:Centrifuge "${params[0]}" called with params: "${JSON.stringify(params[1])}"`)
+}
